+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.1.1]
 
 - Improve documentation
+- Add an empty line between the last `import` and the first `part`
 
 ## [0.1.0]